Make repr module that contains all representations
