+++ CHANGELOG
@@ -3,6 +3,7 @@
 ## Unreleased
 ### Changed
 - Remove style traits and replace all style structs with 3 common ones in `style`
+- Group all representations under a `repr` module.
 
 ## 0.4.0 - 2019-03-02
 ### Added
