+++ examples/boxplot_svg.rs
@@ -1,7 +1,7 @@
 fn main() {
-    let b1 = plotlib::boxplot::BoxPlot::from_slice(&[1.0, 4.0, 2.0, 3.5, 6.4, 2.5, 7.5, 1.8, 9.6])
+    let b1 = plotlib::repr::BoxPlot::from_slice(&[1.0, 4.0, 2.0, 3.5, 6.4, 2.5, 7.5, 1.8, 9.6])
         .label("1");
-    let b2 = plotlib::boxplot::BoxPlot::from_slice(&[3.0, 4.3, 2.0, 3.5, 6.9, 4.5, 7.5, 1.8, 10.6])
+    let b2 = plotlib::repr::BoxPlot::from_slice(&[3.0, 4.3, 2.0, 3.5, 6.9, 4.5, 7.5, 1.8, 10.6])
         .label("2")
         .style(plotlib::style::BoxStyle::new().fill("darkolivegreen"));
     let v = plotlib::view::CategoricalView::new()
