+++ examples/scatter_svg.rs
@@ -9,13 +9,13 @@ fn main() {
         (6.4, 4.3),
         (8.5, 3.7),
     ];
-    let s1 = plotlib::scatter::Scatter::from_slice(&data).style(
+    let s1 = plotlib::repr::Scatter::from_slice(&data).style(
         plotlib::style::PointStyle::new()
             .marker(plotlib::style::PointMarker::Square)
             .colour("burlywood")
             .size(2.),
     );
-    let s2 = plotlib::scatter::Scatter::from_slice(&[(-1.4, 2.5), (7.2, -0.3)])
+    let s2 = plotlib::repr::Scatter::from_slice(&[(-1.4, 2.5), (7.2, -0.3)])
         .style(PointStyle::new().colour("darkseagreen"));
     let v = plotlib::view::ContinuousView::new()
         .add(&s1)
