+++ src/repr/function.rs
@@ -1,40 +1,54 @@
+/*!
+
+Plot arbitrary functions
+
+# Examples
+
+```
+# use plotlib::repr::Function;
+# use plotlib::view::ContinuousView;
+// y=x^2 between 0 and 10
+let f = Function::new(|x| x*x, 0., 10.);
+let v = ContinuousView::new().add(&f);
+```
+*/
+
 use std::f64;
 
 use svg;
 
 use crate::axis;
-use crate::representation::ContinuousRepresentation;
-use crate::style::PointStyle;
+use crate::repr::ContinuousRepresentation;
+use crate::style::LineStyle;
 use crate::svg_render;
-use crate::text_render;
 
-/// The scatter *representation*.
-/// It knows its data as well how to style itself
-#[derive(Debug)]
-pub struct Scatter {
+pub struct Function {
     pub data: Vec<(f64, f64)>,
-    style: PointStyle,
+    style: LineStyle,
 }
 
-impl Scatter {
-    pub fn from_slice(v: &[(f64, f64)]) -> Self {
-        let mut data: Vec<(f64, f64)> = vec![];
-        for &(x, y) in v {
-            data.push((x, y));
-        }
-
-        Scatter {
-            data,
-            style: PointStyle::new(),
+impl Function {
+    pub fn new<F>(f: F, lower: f64, upper: f64) -> Self
+    where
+        F: Fn(f64) -> f64,
+    {
+        let sampling = (upper - lower) / 200.;
+        let samples = (0..)
+            .map(|x| lower + (f64::from(x) * sampling))
+            .take_while(|&x| x <= upper);
+        let values = samples.map(|s| (s, f(s))).collect();
+        Function {
+            data: values,
+            style: LineStyle::new(),
         }
     }
 
-    pub fn style(mut self, style: &PointStyle) -> Self {
+    pub fn style(mut self, style: &LineStyle) -> Self {
         self.style.overlay(style);
         self
     }
 
-    pub fn get_style(&self) -> &PointStyle {
+    pub fn get_style(&self) -> &LineStyle {
         &self.style
     }
 
@@ -59,7 +73,7 @@ impl Scatter {
     }
 }
 
-impl ContinuousRepresentation for Scatter {
+impl ContinuousRepresentation for Function {
     fn range(&self, dim: u32) -> (f64, f64) {
         match dim {
             0 => self.x_range(),
@@ -75,7 +89,7 @@ impl ContinuousRepresentation for Scatter {
         face_width: f64,
         face_height: f64,
     ) -> svg::node::element::Group {
-        svg_render::draw_face_points(
+        svg_render::draw_face_line(
             &self.data,
             x_axis,
             y_axis,
@@ -87,18 +101,11 @@ impl ContinuousRepresentation for Scatter {
 
     fn to_text(
         &self,
-        x_axis: &axis::ContinuousAxis,
-        y_axis: &axis::ContinuousAxis,
-        face_width: u32,
-        face_height: u32,
+        _x_axis: &axis::ContinuousAxis,
+        _y_axis: &axis::ContinuousAxis,
+        _face_width: u32,
+        _face_height: u32,
     ) -> String {
-        text_render::render_face_points(
-            &self.data,
-            x_axis,
-            y_axis,
-            face_width,
-            face_height,
-            &self.style,
-        )
+        "".into()
     }
 }
