+++ src/text_render.rs
@@ -4,7 +4,7 @@ use std;
 use std::collections::HashMap;
 
 use crate::axis;
-use crate::histogram;
+use crate::repr;
 use crate::style;
 use crate::utils::PairWise;
 
@@ -31,7 +31,7 @@ fn tick_offset_map(axis: &axis::ContinuousAxis, face_width: u32) -> HashMap<i32,
 /// and the number of face cells to work with,
 /// create a mapping of cell offset to bin bound
 fn bound_cell_offsets(
-    hist: &histogram::Histogram,
+    hist: &repr::Histogram,
     x_axis: &axis::ContinuousAxis,
     face_width: u32,
 ) -> Vec<i32> {
@@ -271,7 +271,7 @@ pub fn render_x_axis_strings(x_axis: &axis::ContinuousAxis, face_width: u32) -> 
 /// and the face height and width,
 /// create the strings to be drawn as the face
 pub fn render_face_bars(
-    h: &histogram::Histogram,
+    h: &repr::Histogram,
     x_axis: &axis::ContinuousAxis,
     y_axis: &axis::ContinuousAxis,
     face_width: u32,
@@ -591,7 +591,7 @@ mod tests {
     #[test]
     fn test_render_face_bars() {
         let data = vec![0.3, 0.5, 6.4, 5.3, 3.6, 3.6, 3.5, 7.5, 4.0];
-        let h = histogram::Histogram::from_slice(&data, histogram::Bins::Count(10));
+        let h = repr::Histogram::from_slice(&data, repr::HistogramBins::Count(10));
         let x_axis = axis::ContinuousAxis::new(0.3, 7.5);
         let y_axis = axis::ContinuousAxis::new(0., 3.);
         let strings = render_face_bars(&h, &x_axis, &y_axis, 20, 10);
@@ -616,7 +616,7 @@ mod tests {
 
     #[test]
     fn test_render_face_points() {
-        use crate::scatter;
+        use crate::repr;
         use crate::style::PointStyle;
         let data = vec![
             (-3.0, 2.3),
@@ -626,7 +626,7 @@ mod tests {
             (6.4, 4.3),
             (8.5, 3.7),
         ];
-        let s = scatter::Scatter::from_slice(&data);
+        let s = repr::Scatter::from_slice(&data);
         let x_axis = axis::ContinuousAxis::new(-3.575, 9.075);
         let y_axis = axis::ContinuousAxis::new(-1.735, 5.635);
         let style = PointStyle::new();
