+++ tests/test_no_data.rs
@@ -1,5 +1,5 @@
 use plotlib::page::Page;
-use plotlib::scatter::Scatter;
+use plotlib::repr::Scatter;
 use plotlib::style::{PointMarker, PointStyle};
 use plotlib::view::ContinuousView;
